Fix day7/part1
The code to detect who is at the bottom is way simpler than it was
during its development. For every name, we check every holder. As
soon as we pick up a name or a row, we increment the index to avoid
issues. There's a loop to check on holder rows, and another to
check inside the loop every holder. If we reach an empty row, it
means we didn't find any holder, hence we exit. If we find a that
the current name is being held by someone, we go to the very first
loop and pick up another name.