Add day7/part2 (work in progress)
The idea is to have a third array holding the weight of the towers,
rather than the weight of the programs. To fill it in, recursive
calls are made on it until the array is full. For some reason, not
all calls are being made.

New methods are getindex and calctowers (for the weight of them).